backend/pyproject.toml
[project]
name = "hairtryon-backend"
version = "0.1.0"
description = "Hair Try On Backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "alembic>=1.17.1",
    "authlib>=1.6.5",
    "bcrypt<4.1",
    "boto3>=1.40.68",
    "fastapi-mail>=1.5.8",
    "fastapi[standard]>=0.120.4",
    "google-auth>=2.43.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "replicate>=1.0.7",
    "requests>=2.32.5",
    "sqladmin>=0.21.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "coverage>=7.12.0",
    "pre-commit>=4.4.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]
